requirements.txt
altgraph==0.17.4
asgiref==3.10.0
colorama==0.4.6
darkdetect==0.8.0
Django==5.2.7
image==1.5.33
logger==1.4
loguru==0.7.3
MouseInfo==0.1.3
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
psutil==7.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMsgBox==2.0.1
pynput==1.7.7
pyperclip==1.10.0
PyQt6==6.7.1
PyQt6-Fluent-Widgets==1.7.4
PyQt6-Frameless-Window==0.7.3
PyQt6-Qt6==6.7.3
PyQt6_sip==13.10.2
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
pystray==0.19.5
pytweening==1.2.0
pywin32==308
pywin32-ctypes==0.2.3
shiboken6==6.10.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
win32_setctime==1.2.0
